document discovery run and test commands
